backend_project/requirements.txt
fastapi
pydantic
uvicorn
python-multipart
sqlalchemy
aiomysql
cryptography
pymysql
opencv-python
pillow
pytesseract
numpy
